Refactor oauth route
