+++ src/index.ts
@@ -14,12 +14,12 @@ require('dotenv').config({ path: pathToEnvFile });
 import { prepareRedirectURL, prep, pruneRequestBody } from './middleware/prep';
 import { prepareMtUser } from './middleware/user-auth';
 import configureCors from './middleware/cors';
-import { verifyBearerToken, verifyRequestOrigin } from './middleware/auth';
 
 // import initializeDb from './dbs/mt';
 
 import indexRouter from './routes/index';
 import authRouter from './routes/auth';
+import oauthRouter from './routes/oauth';
 
 const configure = (app: express.Application): void => {
   app.set('views', join(__dirname, '../views'));
@@ -34,14 +34,13 @@ const configure = (app: express.Application): void => {
   // middleware
   app.use(configureCors);
   app.use(prep);
-  app.use(verifyBearerToken);
-  app.use(verifyRequestOrigin);
   app.use(prepareMtUser);
   app.use(prepareRedirectURL);
   app.use(pruneRequestBody);
 
   app.use('/', indexRouter);
   app.use('/auth', authRouter);
+  app.use('/oauth', oauthRouter);
 
   // catch 404 and forward to error handler
   app.use(
