+++ src/routes/auth/index.ts
@@ -2,16 +2,18 @@ import express from 'express';
 
 import localLogin from './login';
 import forgot from './forgot';
-import oauth from './oauth';
 import reset from './reset';
 import signup from './signup';
 import accessToken from './accessToken';
+import { verifyBearerToken, verifyRequestOrigin } from '../../middleware/auth';
 
 const router = express.Router();
 
+router.use(verifyBearerToken);
+router.use(verifyRequestOrigin);
+
 router.use('/login', localLogin);
 router.use('/signup', signup);
-router.use('/oauth', oauth);
 router.use('/forgot', forgot);
 router.use('/reset', reset);
 router.use('/accessToken', accessToken);
