+++ src/routes/oauth/google.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+const router = express.Router();
+
+import { googleCallback, googleOauth } from '../../controllers/oauth/google';
+
+router.get('/', googleOauth);
+
+router.get('/callback', googleCallback);
+
+export default router;
